requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyascii==0.3.2
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
datasets==3.1.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
easyocr==1.7.2
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
evaluate==0.4.3
executing==2.1.0
filelock==3.16.1
fire==0.7.0
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
h5py==3.12.1
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imgaug==0.4.0
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jiwer==3.0.5
joblib==1.4.2
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.7
langdetect==1.0.9
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
paddleocr==2.9.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyclipper==1.3.0.post6
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
pypdfium2==4.30.0
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-doctr==0.10.0
python-docx==1.1.2
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.6.0
shapely==2.0.6
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli==2.1.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.0
